ADD: my movements works and rec movements appears
