+++ Beakon/app/src/main/java/erica/beakon/Adapters/MyMovementAdapter.java
@@ -10,29 +10,27 @@ import android.widget.TextView;
 import java.util.ArrayList;
 
 import erica.beakon.MainActivity;
+import erica.beakon.Objects.Movement;
 import erica.beakon.Pages.ExpandedHashtagPage;
 import erica.beakon.R;
 
-/**
- * Created by Cecelia on 11/6/16.
- */
 
-public class MyMovementAdapter extends ArrayAdapter<String> {
+public class MyMovementAdapter extends ArrayAdapter<Movement> {
 
-    public MyMovementAdapter(Context context, ArrayList<String> movements) {
+    public MyMovementAdapter(Context context, ArrayList<Movement> movements) {
         super(context, 0, movements);
     }
 
     @Override
     public View getView(final int position, View convertView, ViewGroup parent) {
-        final String movement = getItem(position);
+        final Movement movement = getItem(position);
 
         if (convertView == null) {
             convertView = LayoutInflater.from(getContext()).inflate(R.layout.my_movement_item, parent, false);
         }
 
         TextView movementNameView = (TextView) convertView.findViewById(R.id.movement_name);
-        movementNameView.setText(movement);
+        movementNameView.setText(movement.getName());
 
         //if hashtag is clicked, go to ExpandedHashtagsPage fragment
         final TextView hashtagsView = (TextView) convertView.findViewById(R.id.hashtag_names);
