--- Beakon/app/src/main/java/erica/beakon/ExpandedMovementsPage.java
@@ -1,30 +1,0 @@
-package erica.beakon;
-
-import android.content.Context;
-import android.net.Uri;
-import android.os.Bundle;
-import android.support.v4.app.Fragment;
-import android.view.LayoutInflater;
-import android.view.View;
-import android.view.ViewGroup;
-
-public class ExpandedMovementsPage extends Fragment {
-
-
-    public ExpandedMovementsPage() {}
-
-    @Override
-    public View onCreateView(LayoutInflater inflater, ViewGroup container,
-                             Bundle savedInstanceState) {
-        // Inflate the layout for this fragment
-        View view = inflater.inflate(R.layout.fragment_expanded_movements_page, container, false);
-
-
-        return view;
-    }
-
-    public interface OnFragmentInteractionListener {
-        // TODO: Update argument type and name
-        void onFragmentInteraction(Uri uri);
-    }
-}
