+++ Beakon/app/src/main/java/erica/beakon/Objects/User.java
@@ -1,5 +1,7 @@
 package erica.beakon.Objects;
 
+import java.util.ArrayList;
+
 /**
  * Created by mafaldaborges on 11/7/16.
  */
@@ -7,6 +9,7 @@ public class User {
     private String id;
     private String name;
     private String email;
+    private ArrayList<String> movements;
 
     public User() {
 
@@ -16,6 +19,7 @@ public class User {
         this.id = id;
         this.name = name;
         this.email = email;
+        this.movements = new ArrayList<>();
     }
 
     public String getId() {
@@ -40,5 +44,21 @@ public class User {
 
     public void setEmail(String email) {
         this.email = email;
+    }
+
+    public ArrayList<String> getMovements() { return this.movements; }
+
+    public void setMovements(ArrayList<String> movements){ this.movements = movements; }
+
+    public void setEmptyMovements() {setMovements(new ArrayList<String>());}
+
+    public void addMovement(Movement movement) { this.movements.add(movement.getId()); }
+
+    public void addMovement(String movementId) { this.movements.add(movementId); }
+
+    public void removeMovement(Movement movement) { this.movements.remove(movement.getId()); }
+
+    public boolean isInMovement(Movement movement) {
+        return this.getMovements().contains(movement.getId());
     }
 }
