+++ Beakon/app/src/main/res/layout/fragment_my_movements_tab.xml
@@ -4,49 +4,55 @@
     android:layout_height="match_parent"
     tools:context=".Pages.MyMovementsTab">
 
-    <LinearLayout
+    <RelativeLayout
         android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:orientation="horizontal"
-        android:id="@+id/tab_buttons">
+        android:layout_height="match_parent">
 
-        <Button
-            android:layout_width="0dp"
-            android:layout_weight="0.5"
-            android:layout_margin="0dp"
-            android:padding="0dp"
+        <LinearLayout
+            android:layout_width="match_parent"
             android:layout_height="wrap_content"
-            android:text="My Movements"
-            android:id="@+id/my_movements"
-            android:background="@color/colorBackground" />
+            android:orientation="horizontal"
+            android:id="@+id/tab_buttons">
 
-        <Button
-            android:layout_width="0dp"
-            android:layout_weight="0.5"
-            android:layout_margin="0dp"
-            android:padding="0dp"
-            android:layout_height="wrap_content"
-            android:text="Suggested"
-            android:id="@+id/movements"
-            android:background="@color/colorAccentLight"/>
+            <Button
+                android:layout_width="0dp"
+                android:layout_weight="0.5"
+                android:layout_margin="0dp"
+                android:padding="0dp"
+                android:layout_height="wrap_content"
+                android:text="My Movements"
+                android:id="@+id/my_movements"
+                android:background="@color/colorBackground" />
 
-    </LinearLayout>
+            <Button
+                android:layout_width="0dp"
+                android:layout_weight="0.5"
+                android:layout_margin="0dp"
+                android:padding="0dp"
+                android:layout_height="wrap_content"
+                android:text="Suggested"
+                android:id="@+id/movements"
+                android:background="@color/colorAccentLight"/>
 
-    <!--<android.support.v4.view.ViewPager-->
-        <!--xmlns:android="http://schemas.android.com/apk/res/android"-->
-        <!--android:id="@+id/pager"-->
-        <!--android:layout_below="@+id/tab_buttons"-->
-        <!--android:layout_width="match_parent"-->
-        <!--android:layout_height="match_parent"-->
-        <!--android:layout_marginLeft="@dimen/activity_horizontal_margin"-->
-        <!--android:layout_marginRight="@dimen/activity_horizontal_margin"-->
-        <!--android:layout_marginTop="@dimen/activity_vertical_margin"-->
-        <!--android:layout_marginBottom="@dimen/activity_vertical_margin"/>-->
+        </LinearLayout>
 
-    <ListView
-        android:layout_width="match_parent"
-        android:layout_height="456dp"
-        android:id="@+id/my_movements_list"
-        android:layout_gravity="center_horizontal|bottom"></ListView>
+        <!--<android.support.v4.view.ViewPager-->
+            <!--xmlns:android="http://schemas.android.com/apk/res/android"-->
+            <!--android:id="@+id/pager"-->
+            <!--android:layout_below="@+id/tab_buttons"-->
+            <!--android:layout_width="match_parent"-->
+            <!--android:layout_height="match_parent"-->
+            <!--android:layout_marginLeft="@dimen/activity_horizontal_margin"-->
+            <!--android:layout_marginRight="@dimen/activity_horizontal_margin"-->
+            <!--android:layout_marginTop="@dimen/activity_vertical_margin"-->
+            <!--android:layout_marginBottom="@dimen/activity_vertical_margin"/>-->
+
+        <ListView
+            android:layout_width="match_parent"
+            android:layout_height="456dp"
+            android:id="@+id/my_movements_list"
+            android:layout_below="@id/tab_buttons"></ListView>
+
+    </RelativeLayout>
 
 </FrameLayout>
