+++ Beakon/app/src/main/res/layout/fragment_recommended_movements.xml
@@ -2,11 +2,45 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="erica.beakon.MyMovementsTab">
+    tools:context="erica.beakon.Pages.RecommendedMovementsTab">
 
-    <ListView
+    <RelativeLayout
         android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:id="@+id/recommended_movements_list"></ListView>
+        android:layout_height="match_parent">
+
+        <LinearLayout
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:orientation="horizontal"
+            android:id="@+id/tab_buttons">
+
+            <Button
+                android:layout_width="0dp"
+                android:layout_weight="0.5"
+                android:layout_margin="0dp"
+                android:padding="0dp"
+                android:layout_height="wrap_content"
+                android:text="My Movements"
+                android:id="@+id/my_movements"
+                android:background="@color/colorAccentLight" />
+
+            <Button
+                android:layout_width="0dp"
+                android:layout_weight="0.5"
+                android:layout_margin="0dp"
+                android:padding="0dp"
+                android:layout_height="wrap_content"
+                android:text="Suggested"
+                android:id="@+id/movements"
+                android:background="@color/colorBackground"/>
+        </LinearLayout>
+
+        <ListView
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:id="@+id/recommended_movements_list"></ListView>
+
+    </RelativeLayout>
+
 
 </FrameLayout>
